Fixed a show stopping error.
Corrected cast on comparisons between ERF pointers and reclen. This
was causing the acquire function to break out every time.

+++ daq_endace.c
@@ -179,7 +179,7 @@ static int endace_daq_acquire(void *handle, int cnt, DAQ_Analysis_Func_t callbac
 			reclen = ntohs(rec->rlen);
 
 			/*  Advance the stream if a short read is detected */
-			if ((uint8_t)(ep-cp) < reclen) break;
+			if ((ep-cp) < reclen) break;
 
 			/* Advance the current pointer */
 			cp += reclen;
